move definitions from Debug.hpp
